+++ packages/addresses-tools/src/addresses.helpers.ts
@@ -12,10 +12,14 @@ import {
   MainnetAddressGroup,
   MainnetNetworkName,
   NetworkAddresses,
+  NetworkName,
   TestnetAddressGroup,
   TestnetNetworkName,
   ZetaAddress,
+  ZetaLocalNetworkName,
+  ZetaMainnetNetworkName,
   ZetaNetworkName,
+  ZetaTestnetNetworkName,
 } from "@zetachain/addresses";
 import dotenv from "dotenv";
 import { readdirSync, readFileSync, writeFileSync } from "fs";
@@ -184,4 +188,57 @@ export const addNewNetwork = (newNetworkName: string, addTo: ZetaNetworkName[]) 
   });
 
   console.log(`To enable IntelliSense, add the network (${newNetworkName}) to the constants (addresses.helpers.ts).`);
+};
+
+export const getLocalnetListAsync = async (): Promise<Record<ZetaLocalNetworkName, LocalnetAddressGroup>> => {
+  const troy = await require("./addresses.troy.json");
+  return {
+    troy: troy as LocalnetAddressGroup,
+  };
+};
+
+export const getTestnetListAsync = async (): Promise<Record<ZetaTestnetNetworkName, TestnetAddressGroup>> => {
+  const athens = await require("./addresses.athens.json");
+  return {
+    athens: athens as TestnetAddressGroup,
+  };
+};
+
+export const getMainnetListAsync = async (): Promise<Record<ZetaMainnetNetworkName, MainnetAddressGroup>> => {
+  const mainnet = await require("./addresses.mainnet.json");
+  return {
+    mainnet: mainnet as MainnetAddressGroup,
+  };
+};
+
+export const loadAddressFromFile = async (
+  address: ZetaAddress,
+  {
+    customNetworkName,
+    customZetaNetwork,
+  }: { customNetworkName?: NetworkName; customZetaNetwork?: ZetaNetworkName } = {}
+): Promise<string> => {
+  const { name: _networkName } = network;
+  const networkName = customNetworkName || _networkName;
+
+  const { ZETA_NETWORK: _ZETA_NETWORK } = process.env;
+  const ZETA_NETWORK = customZetaNetwork || _ZETA_NETWORK;
+
+  if (!ZETA_NETWORK) throw MissingZetaNetworkError;
+
+  console.log(`Getting ${address} address from ${ZETA_NETWORK}: ${networkName}.`);
+
+  if (isZetaLocalnet(ZETA_NETWORK) && isLocalNetworkName(networkName)) {
+    return (await getLocalnetListAsync())[ZETA_NETWORK][networkName][address];
+  }
+
+  if (isZetaTestnet(ZETA_NETWORK) && isTestnetNetworkName(networkName)) {
+    return (await getTestnetListAsync())[ZETA_NETWORK][networkName][address];
+  }
+
+  if (isZetaMainnet(ZETA_NETWORK) && isMainnetNetworkName(networkName)) {
+    return (await getMainnetListAsync())[ZETA_NETWORK][networkName][address];
+  }
+
+  throw new Error(`Invalid ZETA_NETWORK + network combination ${ZETA_NETWORK} ${networkName}.`);
 };
